fix: Implementa correção para deletar uma conta bancaria

+++ src/main/java/com/vcontrola/vcontrola/service/ContaService.java
@@ -7,7 +7,9 @@ import com.vcontrola.vcontrola.entity.TipoContaUsuario;
 import com.vcontrola.vcontrola.entity.Usuario;
 import com.vcontrola.vcontrola.mapper.ContaMapper;
 import com.vcontrola.vcontrola.repository.ContaRepository;
+import com.vcontrola.vcontrola.repository.ItemPlanejamentoRepository;
 import com.vcontrola.vcontrola.repository.TipoContaUsuarioRepository;
+import com.vcontrola.vcontrola.repository.TransacaoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,15 +21,18 @@ public class ContaService {
 
     private final ContaRepository repository;
 
-
+    private  final TransacaoRepository transacaoRepository;
     private final ContaMapper mapper;
 
     private final TipoContaUsuarioRepository tipoRepository;
+    private final ItemPlanejamentoRepository itemPlanejamentoRepository;
 
-    public ContaService(ContaRepository repository, ContaMapper mapper, TipoContaUsuarioRepository tipoRepository) {
+    public ContaService(ContaRepository repository, TransacaoRepository transacaoRepository, ContaMapper mapper, TipoContaUsuarioRepository tipoRepository, ItemPlanejamentoRepository itemPlanejamentoRepository) {
         this.repository = repository;
+        this.transacaoRepository = transacaoRepository;
         this.mapper = mapper;
         this.tipoRepository = tipoRepository;
+        this.itemPlanejamentoRepository = itemPlanejamentoRepository;
     }
 
     public void criar(ContaRequest dados, Usuario usuario) {
@@ -69,12 +74,26 @@ public class ContaService {
     }
 
     public void excluir(UUID id, Usuario usuario) {
+
+
         Conta conta = repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Conta não encontrada"));
+                .orElseThrow(() -> new RuntimeException("Conta não encontrada."));
+
 
         if (!conta.getUsuario().getId().equals(usuario.getId())) {
-            throw new RuntimeException("Você não tem permissão para excluir esta conta.");
+            throw new RuntimeException("Acesso negado: Você não pode excluir esta conta.");
         }
+
+
+        if (transacaoRepository.existsByContaId(id)) {
+            throw new RuntimeException("Não é possível excluir: Existem transações vinculadas a esta conta. Exclua as transações primeiro.");
+        }
+
+
+        if (itemPlanejamentoRepository.existsByContaDestinoId(id)) {
+            throw new RuntimeException("Não é possível excluir: Esta conta está sendo usada no Planejamento Financeiro.");
+        }
+
 
         repository.delete(conta);
     }
